Move SwapWorkflowStep enum to separate file and rename.
