+++ src/main/java/swaps/view/SwapWorkflowStep.java
@@ -1,0 +1,18 @@
+package swaps.view;
+
+public enum SwapWorkflowStep {
+    SELECT_GOAL_TYPE("Select Goal Type"),
+    PRECISE_GOAL_DETAILS("Precise Goal Details"),
+    IMPRECISE_GOAL_DETAILS("Imprecise Goal Details");
+
+    private final String label;
+
+    SwapWorkflowStep(String label) {
+        this.label = label;
+    }
+
+    @Override
+    public String toString() {
+        return label;
+    }
+}
